+++ src/ipc/types.js
@@ -131,6 +131,7 @@
  * @typedef {object} EngineStateChangedPayload
  * @property {"running"|"stopped"|"error"} state
  * @property {string|undefined} error
+ * @property {"deviceInvalidated"|undefined} reason
  */
 
 /**